Fix ARM CE byte ordering, expand C/C++ API, and harden build
ARM backends: fix round key byte-swap on little-endian (vrev32q_u8),
rewrite decrypt to pre-process middle keys with InvMixColumns, fix
GHASH PMULL reflect and reduction ordering.

API: add nonce/IV-generating convenience overloads for CTR, CBC, and
GCM (library generates and prepends nonce, appends tag). Add C API
for IV/nonce generation. Rename error codes (TINYAES_OK, Result::Ok,
Result::AuthenticationFailed, etc.).

Build: add MinGW GCC AVX-512 debug alignment fix, harden bench/fuzz
CMake targets (warnings-as-errors, linker hardening), align with
tinysha CMake conventions. Add README.

Tests: expand coverage for nonce-generating API overloads, add NIST
GCM test vectors, improve fuzz target differential testing.

+++ tests/test_harness.h
@@ -75,10 +75,10 @@ namespace test
         {
             std::fprintf(stderr, "%s:%d: ASSERT_EQ failed\n  got:      ", file, line);
             for (auto x : a)
-                std::fprintf(stderr, "%02x", x);
+                std::fprintf(stderr, "%02x", static_cast<unsigned>(x));
             std::fprintf(stderr, "\n  expected: ");
             for (auto x : b)
-                std::fprintf(stderr, "%02x", x);
+                std::fprintf(stderr, "%02x", static_cast<unsigned>(x));
             std::fprintf(stderr, "\n");
             fail_count()++;
         }
